backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
requests
rembg
opencv-python
pytesseract
torch
psutil
onnxruntime
pandas